# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NomadCrew RESTful API with authentication and WebSocket support",
        "title": "NomadCrew Backend API",
        "contact": {
            "name": "NomadCrew Team",
            "url": "https://nomadcrew.uk",
            "email": "support@nomadcrew.uk"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "api.nomadcrew.uk",
    "basePath": "/v1",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pagination offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination limit (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users with pagination info",
                        "schema": {
                            "$ref": "#/definitions/docs.UserListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid pagination parameters",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the profile of the currently authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get current user profile",
                "responses": {
                    "200": {
                        "description": "User profile",
                        "schema": {
                            "$ref": "#/definitions/types.UserProfile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - No authenticated user",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a user profile by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile",
                        "schema": {
                            "$ref": "#/definitions/types.UserProfile"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid user ID format",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - User not found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates fields for a specified user. Only admin or the user themselves can perform this action.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.UserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated user profile",
                        "schema": {
                            "$ref": "#/definitions/types.UserProfile"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid user ID or request body",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User not authorized to update this profile",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - User not found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/preferences": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the preferences for a specified user. Only admin or the user themselves can perform this action.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user preferences",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID whose preferences to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User preferences map",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.UserPreferencesRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Preferences updated successfully"
                    },
                    "400": {
                        "description": "Bad request - Invalid user ID or preferences format",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User not authorized to update these preferences",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - User not found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.ErrorResponse": {
            "description": "Error information",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "error": {
                    "description": "Detailed error information",
                    "type": "string",
                    "example": "Field 'name' is required"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Invalid request parameters"
                }
            }
        },
...
        "docs.UserListResponse": {
            "description": "A paginated list of user profiles.",
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.UserProfile"
                    }
                }
            }
        },
        "docs.UserPreferencesRequest": {
            "description": "A map of preference keys to their values.",
            "type": "object",
            "additionalProperties": true
        },
        "docs.UserUpdateRequest": {
            "description": "Fields for updating a user profile. All fields are optional.",
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string",
                    "example": "Jonathan"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doer"
                },
                "profilePictureUrl": {
                    "type": "string",
                    "example": "https://example.com/new_avatar.png"
                },
                "username": {
                    "type": "string",
                    "example": "new.john.doe"
                }
            }
        },
...
        "types.UserProfile": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isOnline": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "lastSeenAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}